cleaned up google stories
