+++ packages/lib/src/components/GooglePay/stories/GooglePayExpressSessionsDemo.tsx
@@ -6,166 +6,13 @@ import { createSession, patchCheckoutSession } from '../../../../storybook/helpe
 import { RETURN_URL, STORYBOOK_ENVIRONMENT_URLS } from '../../../../storybook/config/commonConfig';
 import GooglePay from '../GooglePay';
 import { AdyenCheckout } from '../../../core/AdyenCheckout';
-
-import type { GooglePayConfiguration } from '../types';
-import type { MetaConfiguration, StoryConfiguration } from '../../../../storybook/types';
+import { createGooglePayAmountHelper, getShippingOptions, getTransactionInfo, EXPRESS_DEMO_SETTINGS } from './googlePayExpressUtils';
 import type { PaymentAmount } from '../../../types/global-types';
 import type { CheckoutSession, CoreConfiguration } from '../../../types';
 
-type GooglePayStory = StoryConfiguration<GooglePayConfiguration>;
+const GooglePayAmountHelper = createGooglePayAmountHelper(EXPRESS_DEMO_SETTINGS.INITIAL_AMOUNT);
 
-const COUNTRY_CODE = 'BR';
-const SHOPPER_LOCALE = 'en-US';
-const INITIAL_AMOUNT = 10000;
-
-let finalAmount = INITIAL_AMOUNT;
-
-const meta: MetaConfiguration<GooglePayConfiguration> = {
-    title: 'Components/Wallets/GooglePay',
-    tags: ['no-automated-visual-test']
-};
-
-type ShippingCosts = Record<string, string>;
-
-/**
- * Method that calculate the shipping costs based on the country/shipping options
- *
- * @param countryCode - country code
- */
-function getShippingCostByCountry(countryCode: string): ShippingCosts {
-    switch (countryCode) {
-        case 'BR':
-            return {
-                'shipping-001': '8.00',
-                'shipping-002': '25.00'
-            };
-        default: {
-            return {
-                'shipping-001': '0.00',
-                'shipping-002': '5.00',
-                'shipping-003': '15.00'
-            };
-        }
-    }
-}
-
-/**
- * Method that fetches the shipping options according to the country.
- * This function in most of the cases is asynchronous, as it will request shipping options in the backend
- *
- * @param countryCode - country code
- */
-function getShippingOptions(countryCode?: string): Promise<google.payments.api.ShippingOptionParameters> {
-    switch (countryCode) {
-        case 'BR': {
-            return Promise.resolve({
-                defaultSelectedOptionId: 'shipping-001',
-                shippingOptions: [
-                    {
-                        id: 'shipping-001',
-                        label: 'R$ 8.00: Standard shipping',
-                        description: 'Free Shipping delivered in 10 business days.'
-                    },
-                    {
-                        id: 'shipping-002',
-                        label: 'R$ 25.00: Super Express shipping',
-                        description: 'Standard shipping delivered in 2 business days.'
-                    }
-                ]
-            });
-        }
-        default: {
-            return Promise.resolve({
-                defaultSelectedOptionId: 'shipping-001',
-                shippingOptions: [
-                    {
-                        id: 'shipping-001',
-                        label: 'R$ 0.00: Free shipping',
-                        description: 'Free Shipping delivered in 10 business days.'
-                    },
-                    {
-                        id: 'shipping-002',
-                        label: 'R$ 5.00: Fedex shipping',
-                        description: 'Standard shipping delivered in 5 business days.'
-                    },
-                    {
-                        id: 'shipping-003',
-                        label: 'R$ 15.00: DHL Express shipping',
-                        description: 'Express shipping delivered in 2 business day.'
-                    }
-                ]
-            });
-        }
-    }
-}
-
-function getTransactionInfo(): google.payments.api.TransactionInfo {
-    return {
-        displayItems: [
-            {
-                label: 'Subtotal',
-                type: 'SUBTOTAL',
-                price: '99.00'
-            },
-            {
-                label: 'Tax',
-                type: 'TAX',
-                price: '1.00'
-            }
-        ],
-        countryCode: 'BR',
-        currencyCode: 'BRL',
-        totalPriceStatus: 'FINAL',
-        totalPrice: '100.00',
-        totalPriceLabel: 'Total'
-    };
-}
-
-function convertFloatAmountToAdyenAmount(totalPrice: string): number {
-    if (totalPrice.includes('.')) {
-        return Number(totalPrice.replace('.', ''));
-    }
-    return Number(`${totalPrice}00`);
-}
-
-function calculateNewTransactionInfo(countryCode: string, selectedShippingOptionId: string) {
-    const newTransactionInfo = getTransactionInfo();
-    const shippingCost = getShippingCostByCountry(countryCode)[selectedShippingOptionId] ?? '0.00';
-
-    newTransactionInfo.displayItems?.push({
-        type: 'LINE_ITEM',
-        label: 'Shipping cost',
-        price: shippingCost,
-        status: 'FINAL'
-    });
-
-    let totalPrice = 0.0;
-    newTransactionInfo.displayItems?.forEach(displayItem => (totalPrice += parseFloat(displayItem.price)));
-    newTransactionInfo.totalPrice = totalPrice.toString();
-
-    finalAmount = convertFloatAmountToAdyenAmount(totalPrice.toString());
-
-    return newTransactionInfo;
-}
-
-export const ExpressOnSessions: GooglePayStory = {
-    render: checkoutConfig => {
-        const { amount, countryCode, shopperLocale } = checkoutConfig;
-        return <GooglePayExpressSessionsDemo amount={amount} countryCode={countryCode} shopperLocale={shopperLocale} />;
-    },
-    parameters: {
-        controls: { exclude: ['useSessions', 'shopperLocale', 'amount', 'showPayButton', 'countryCode', 'srConfig.showPanel'] }
-    },
-    args: {
-        countryCode: COUNTRY_CODE,
-        amount: INITIAL_AMOUNT,
-        shopperLocale: SHOPPER_LOCALE
-    }
-};
-
-export default meta;
-
-function InfoBox() {
+export function InfoBox() {
     return (
         <div
             style={{
@@ -192,7 +39,7 @@ function InfoBox() {
     );
 }
 
-function GooglePayExpressSessionsDemo({
+export function GooglePayExpressSessionsDemo({
     amount,
     countryCode,
     shopperLocale
@@ -252,8 +99,8 @@ function GooglePayExpressSessionsDemo({
                     const { session } = component.core.session;
 
                     const amountToBePatched = {
-                        currency: getCurrency(COUNTRY_CODE),
-                        value: finalAmount
+                        currency: getCurrency(EXPRESS_DEMO_SETTINGS.COUNTRY_CODE),
+                        value: GooglePayAmountHelper.getFinalAmount()
                     };
 
                     const sessionData = await patchSessionAndMakePayable(amountToBePatched, session);
@@ -295,7 +142,7 @@ function GooglePayExpressSessionsDemo({
 
             paymentDataCallbacks: {
                 onPaymentDataChanged(intermediatePaymentData) {
-                    // eslint-disable-next-line no-async-promise-executor
+                    // eslint-disable-next-line no-async-promise-executor, @typescript-eslint/no-misused-promises
                     return new Promise(async resolve => {
                         const { callbackTrigger, shippingAddress, shippingOptionData } = intermediatePaymentData;
                         const paymentDataRequestUpdate: google.payments.api.PaymentDataRequestUpdate = {};
@@ -317,7 +164,7 @@ function GooglePayExpressSessionsDemo({
                                     paymentDataRequestUpdate.newShippingOptionParameters?.defaultSelectedOptionId ||
                                     paymentDataRequestUpdate.newShippingOptionParameters.shippingOptions[0].id;
 
-                                paymentDataRequestUpdate.newTransactionInfo = calculateNewTransactionInfo(
+                                paymentDataRequestUpdate.newTransactionInfo = GooglePayAmountHelper.calculateNewTransactionInfo(
                                     shippingAddress.countryCode,
                                     selectedShippingOptionId
                                 );
@@ -326,7 +173,7 @@ function GooglePayExpressSessionsDemo({
                         /** If SHIPPING_OPTION changed, we calculate the new fee */
                         if (callbackTrigger === 'SHIPPING_OPTION') {
                             if (shippingAddress?.countryCode && shippingOptionData?.id) {
-                                paymentDataRequestUpdate.newTransactionInfo = calculateNewTransactionInfo(
+                                paymentDataRequestUpdate.newTransactionInfo = GooglePayAmountHelper.calculateNewTransactionInfo(
                                     shippingAddress.countryCode,
                                     shippingOptionData.id
                                 );
